rpa_conciliaciones/requirements.txt
customtkinter>=5.2
pyautogui>=0.9.54
pynput>=1.7
pillow>=10.0
pyperclip>=1.8
pygetwindow>=0.0.9
pandas>=2.0
openpyxl>=3.1
httpx>=0.27
keyring>=25.0
pyinstaller>=6.0
python-dateutil>=2.9
semver>=3.0
tkcalendar>=1.6
